Add cautionary note to playground instructions in Chinese guide
A new line of instruction is added to the 'playground' section of the Chinese guide. This line provides a cautionary note to alert users that the playground will reset after page refresh, advising them not to use it for continuous editing. This additional information seeks to prevent potential confusion or loss of data.