Add markdown table tutorials to step-6
Updated the content of `step-6` in the tutorial to include instructions on how to create and format tables in Markdown. This involves information on how to use vertical bars and hyphens to divide columns and headers, alignment adjustments, as well as practical exercises for learners to create their own tables. The 'hint' section and the app templates were also updated correspondingly.