Add Chinese Markdown cheatsheet and update navigation
A new Markdown cheatsheet resource for Chinese users has been added. This provides an easy reference for the Markdown language syntax. In addition to this, the tutorial and reference sections in the Chinese section of the website navigation have been updated and fixed to correctly link to the right resources.